mhd-nlp-docker/requirements.txt
pymongo
python-multipart
python-dotenv
python-jose
PyMuPDF
pdfminer.six
uvicorn[standard]
fastapi
spacy==3.7.5
scispacy==0.5.5
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
authlib>=1.3.0
httpx
itsdangerous>=2.1.2
pytesseract==0.3.13
pdf2image==1.17.0
pillow>=10.0.0
PyPDF2==3.0.1
pydantic>=1.10
python-dateutil
cryptography
motor
minio
boto3
azure-storage-blob
snowflake-connector-python
mlflow==2.16.0
scikit-learn
pandas
numpy
sqlalchemy
redis==5.0.8
rq==1.15.1
en-core-web-sm==3.7.1